+++ src/main/java/com/springbootstudy/app/service/NoticeBoardService.java
@@ -2,6 +2,7 @@ package com.springbootstudy.app.service;
 
 import java.util.List;
 
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.springbootstudy.app.domain.NoticeBoard;
@@ -10,25 +11,35 @@ import com.springbootstudy.app.mapper.NoticeBoardMapper;
 import lombok.extern.slf4j.Slf4j;
 
 @Service
-public interface NoticeBoardService {
+@Slf4j
+public class NoticeBoardService {
+	
+	@Autowired
+	private NoticeBoardMapper noticeBoardMapper;
 
 	// 현재 페이지 글 리스트 반환
-	List<NoticeBoard> noticeBoardList();
+	public List<NoticeBoard> noticeBoardList(){
+		return noticeBoardMapper.noticeBoardList();
+	}
 	
 	// no에 해달하는 게시글 반환
-    NoticeBoard getNoticeBoard(int no);
+    public NoticeBoard getNoticeBoard(int no) {
+    	return noticeBoardMapper.getNoticeBoard(no);
+    }
     
     // 게시글 추가
-    void addNoticeBoard(NoticeBoard noticeBoard);
-    
-    
-    // 비밀번호 체크
-    public boolean isPassCheck(int no, String pass);
+    public void addNoticeBoard(NoticeBoard noticeBoard) {
+    	noticeBoardMapper.addNoticeBoard(noticeBoard);
+    }
     
     // 게시글 수정
-    void updateNoticeBoard(NoticeBoard noticeBoard);
+    public void updateNoticeBoard(NoticeBoard noticeBoard) {
+    	noticeBoardMapper.updateNoticeBoard(noticeBoard);
+    }
     
     // 게시글 삭제
-    void deleteNoticeBoard(int no);
+    public void deleteNoticeBoard(int no) {
+    	noticeBoardMapper.deleteNoticeBoard(no);
+    }
    
 }
