+++ src/main/resources/static/css/noticeList.css
@@ -1,0 +1,75 @@
+@import
+	url('https://fonts.googleapis.com/css2?family=Noto+Sans+KR&display=swap');
+
+#global-content {
+	font-family: 'Noto Sans KR', sans-serif;
+	margin: 50px 0;
+}
+
+.notice-header {
+	margin-bottom: 30px;
+}
+
+.notice-header h2 {
+	font-size: 28px;
+	font-weight: bold;
+}
+
+.write-button {
+	display: inline-block;
+	padding: 6px 14px;
+	background-color: transparent;
+	color: #7a8577;
+	text-decoration: none;
+	border-radius: 5px;
+	margin-bottom: 20px;
+	border: 1px solid #7a8577;
+	transition: all 0.3s ease;
+	font-size: 0.9em;
+}
+
+.write-button:hover {
+	background-color: #7a8577;
+	color: white;
+	border-color: #7a8577;
+}
+
+.notice-table {
+	width: 100%;
+	border-collapse: collapse;
+}
+
+.notice-table thead th {
+	background-color: #f8f9fa;
+	padding: 10px;
+	text-align: center;
+	border-bottom: 2px solid #FF914D;
+}
+
+.notice-table tbody td {
+	padding: 10px;
+	text-align: center;
+	background-color: white;
+	border-top: 1px solid #dee2e6;
+	border-bottom: 1px solid #dee2e6;
+}
+
+.notice-table tbody tr:hover td {
+	/* background-color: #E6EBE3; */
+	background-color: #F7F4F1;
+}
+
+.notice-table .notice-title {
+	text-align: left;
+}
+
+.notice-table .notice-title a {
+	color: inherit;
+	text-decoration: none;
+}
+
+.no-notices {
+	text-align: center;
+	padding: 50px 0;
+	color: #6c757d;
+}
