+++ src/main/resources/templates/fragments/header.html
@@ -4,6 +4,7 @@
   <meta charset="UTF-8">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <title>YOU CAN DO EAT</title>
+  <link rel="icon" href="images/bowl-food-solid.png">
   <style>
   </style> 
   <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-QWTKZyjpPEjISv5WaRU9OFeRpok6YctnYmDr5pNlyT2bRjXh0JMhjY6hW+ALEwIH" crossorigin="anonymous"> 
@@ -13,7 +14,7 @@
     <div class="headerTop"></div>
     <div class="mainLogo"> 
       <div class="imgo">
-      <img src="images/YouCanDoEat.png" alt="logo" class="logo">
+      <a href="/"><img src="images/YouCanDoEat.png" alt="logo" class="logo"></a>
       </div>
       <div class="loginJoin">
         <a class="login" href="#">로그인&nbsp;&nbsp; </a>
@@ -25,9 +26,9 @@
   <!-- nav바 영역-->
    <nav>
         <div class="dropdown">
-            <button class="dropbtn">YOU CAN DO EAT</button>
+            <a href="/" class="dropbtn">YOU CAN DO EAT</a>
             <div class="dropdown-content">
-                <a href="/">한식</a>
+                <a href="#">한식</a>
                 <a href="#">중식</a>
                 <a href="#">일식</a>
                 <a href="#">양식</a>
@@ -36,16 +37,16 @@
         </div>
 
         <div class="dropdown">
-            <button class="dropbtn">공지 & 이벤트</button>
+            <a href="/noticeList" class="dropbtn">공지 & 이벤트</a>
             <div class="dropdown-content">
                 <a href="/noticeList">공지사항</a>
                 <a href="#">이벤트</a>
-                <a href="#">혹시몰라 두는</a>
+                <a href="#">FAQ</a>
             </div>
         </div>
 
         <div class="dropdown">
-            <button class="dropbtn">카테고리 3</button>
+            <a href="#" class="dropbtn">커뮤니티</a>
             <div class="dropdown-content">
                 <a href="#">항목 3-1</a>
                 <a href="#">항목 3-2</a>
