+++ src/main/resources/templates/layouts/main_layout.html
@@ -4,11 +4,11 @@
   <head>
     <meta charset="utf-8">
     <meta name="viewport" content="width=device-width, initial-scale=1">
-    <title>BoardService</title>
+    <title>YOUCANDOEAT</title>
+     <link rel="icon" href="images/utensils-solid.png">
     <link th:href="@{bootstrap/bootstrap.min.css}" rel="stylesheet" >
 	<link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.11.3/font/bootstrap-icons.min.css">
 	<script src="js/jquery-3.2.1.min.js"></script>
-	<script src="js/comment.js"></script>
 	<script src="js/recipeBook.js"></script>
 	<script src="js/recipeList.js"></script>
 	<link href="css/header.css" rel="stylesheet">
