+++ src/main/resources/templates/layouts/notice_layout.html
@@ -5,14 +5,14 @@
     <meta charset="utf-8">
     <meta name="viewport" content="width=device-width, initial-scale=1">
     <title>YOUCANDOEAT</title>
+     <link rel="icon" href="images/utensils-solid.png">
     <link th:href="@{bootstrap/bootstrap.min.css}" rel="stylesheet" >
 	<link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.11.3/font/bootstrap-icons.min.css">
 	<script th:src="@{js/jquery-3.2.1.min.js}"></script>
  	<script th:src="@{js/formcheck.js}"></script>
  	<link href="css/header.css" rel="stylesheet">
 	<link href="css/footer.css" rel="stylesheet">
-	
-    <style>      	
+    <style> 
     </style>
   </head>
   <body>
@@ -20,11 +20,11 @@
   		<!-- 헤더 -->
   	  <div th:replace="~{fragments/header::header}"></div> 
 		
-		<!--/* content : 요청 마다 변경되는 부분 */-->
+		<!--/* content */-->
 		<div layout:fragment="content"></div>
 		
 		<!-- 푸터 -->
-	<div th:replace="~{fragments/footer::footer}"></div> 
+		<div th:replace="~{fragments/footer::footer}"></div> 
   		
 	</div>
     <script src="bootstrap/bootstrap.bundle.min.js"></script>
