+++ src/main/resources/templates/views/noticeList.html
@@ -1,78 +1,53 @@
 <!DOCTYPE html>
-<!--/* 
-	Thymeleaf와 Thymeleaf Layout을 사용하기 위한 NameSpace 정의
-	layout:decorate 옵션은 아래의 <th:block layout:fragment="content"> 부분을 어떤
-	레이아웃에 적용할지 설정하는 부분으로 경로 지정은 templates를 기준으로 지정하면 됨
-*/-->
 <html xmlns:th="http://www.thymeleaf.org"
-		xmlns:layout="http://www.ultraq.net.nz/thymeleaf/layout"		
-		layout:decorate="~{layouts/notice_layout}">		
-	<!--/* content */-->
-	<!--/* 
-		block을 사용해도 되고 div에 직접 layout:fragment를 적용해도 됨
-		main_layout.html에서 지정한 layout:fragment의 이름과 같아야 함  
-	*/-->
+	xmlns:layout="http://www.ultraq.net.nz/thymeleaf/layout"
+	layout:decorate="~{layouts/notice_layout}">
+<head>
+<link rel="stylesheet" type="text/css" th:href="@{/css/noticeList.css}" />
+</head>
+
+<body>
 	<th:block layout:fragment="content">
-		<div class="row my-5" id="global-content">
-			<div class="col-10 offset-1">
-				<div class="row text-center">
-					<div class="col">
-						<h2 class="fs-3 fw-bold">공지사항</h2>
-					</div>
-				</div>  		
-			<!-- 	<form name="searchForm" id="searchForm" action="noticeBoardList" 
-					class="row justify-content-center my-3">
-					<div class="col-auto">
-						<select name="type" class="form-select">
-							<option value="title">제목</option>
-							<option value="writer">작성자</option>
-							<option value="content">내용</option>
-						</select>
-					</div>
-					<div class="col-4">
-						<input type="text" name="keyword" class="form-control"/>
-					</div>
-					<div class="col-auto">
-						<input type="submit" value="검 색" class="btn btn-primary"/>
-					</div>
-				</form>  --> 		
-				<div class="row">
-					<div class="col text-end">
-						<a th:href="@{addNoticeBoard}" class="btn btn-outline-success">글쓰기</a>
-					</div>
-				</div>
-				<div class="row my-3">  			
-					<div class="col">
-						<table class="table">
-							<thead>
-								<tr class="table-dark">
-									<th>NO</th>
-									<th>제목</th>
-									<th>작성자</th>
-									<th>작성일</th>
-									<th>조회수</th>
-								</tr>
-							</thead>								
-							<tbody>
-								<th:block th:if="${not #lists.isEmpty(noticeList)}">										
-									<tr th:block th:each="notice, status: ${noticeList}">
-										<td>[[ ${notice.no} ]]</td>
-										<td><a th:href="@{noticeDetail(no=${notice.no})}">[[ ${notice.title} ]]</a></td>
-										<td th:text="${ notice.writer }"></td>
-										<td>[[${ #dates.format(notice.regDate, 'yyyy-MM-dd') }]]</td>
-										<td th:text="${ notice.readCount }"></td>
-									</tr>										
-								</th:block>
-								<th:block th:unless="${not #lists.isEmpty(noticeList)}">
-									<tr>
-										<td colspan="5">공지사항이 존재하지 않습니다😅</td>											
-									</tr>
-								</th:block>
-							</tbody>								
-						</table>
-					</div>  			
-				</div>
+		<div id="global-content">
+			<div class="notice-header text-center">
+				<h2>공지사항</h2>
 			</div>
+
+			<div class="text-end">
+				<a th:href="@{addNoticeBoard}" class="write-button">공지 글쓰기</a>
+			</div>
+
+			<div class="notice-table-wrapper">
+				<table class="notice-table">
+					<thead>
+						<tr>
+							<th>번&nbsp;호</th>
+							<th class="notice-title text-center">제&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;목</th>
+							<th>작성일</th>
+							<th>조회수</th>
+						</tr>
+					</thead>
+					<tbody>
+						<th:block th:if="${not #lists.isEmpty(noticeList)}">
+							<tr th:each="notice, status: ${noticeList}">
+								<td th:text="${notice.no}"></td>
+								<td class="notice-title"><a
+									th:href="@{noticeDetail(no=${notice.no})}"
+									th:text="${notice.title}"></a></td>
+								<td th:text="${#dates.format(notice.regDate, 'yyyy-MM-dd')}"></td>
+								<td th:text="${notice.readCount}"></td>
+							</tr>
+						</th:block>
+						<th:block th:unless="${not #lists.isEmpty(noticeList)}">
+							<tr>
+								<td colspan="4" class="no-notices">공지사항이 존재하지 않습니다😅</td>
+							</tr>
+						</th:block>
+					</tbody>
+				</table>
+			</div>
+
 		</div>
 	</th:block>
-</html>	
+</body>
+</html>
